Refactor and cleanup codebase
