+++ .gitignore
@@ -1,2 +1,3 @@
+.idea/
 *.pyc
 __pycache__
